requirements.txt
anndata==0.10.6
array_api_compat==1.4.1
contourpy==1.2.0
cycler==0.12.1
Cython==0.29.37
decorator==5.1.1
Deprecated==1.2.14
fonttools==4.49.0
future==1.0.0
graphtools==1.5.3
h5py==3.10.0
hdbscan==0.8.33
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.3
natsort==8.4.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
packaging==24.0
pandas==2.0.3
patsy==0.5.6
phate==1.0.11
pillow==10.2.0
polars==0.20.15
PyGSP==0.5.1
pynndescent==0.5.11
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
s-gd2==1.8.1
scanpy==1.9.8
scikit-learn==1.4.1.post1
scikit-spatial==7.2.0
scipy==1.12.0
scprep==1.2.3
seaborn==0.13.2
session-info==1.0.0
six==1.16.0
statsmodels==0.14.1
stdlib-list==0.10.0
tasklogger==1.2.0
threadpoolctl==3.3.0
tqdm==4.66.2
tzdata==2024.1
umap-learn==0.5.5
wrapt==1.16.0
